requirements.txt
mistral_inference
huggingface_hub
joblib

pandas
pydantic
pyarrow
fastparquet
ruamel.yaml